Renamed gcp-compute-service-adapter to gcp-load-balancer
